diagnostics/virt: preserve DMI read-failure path, fix stale docs, unify with osutils::virt
Three fixes from Copilot review on PR778:

- diagnostics::get_virtualization_info collapsed vendor/product DMI read
  failures into one generic "virtualization info" label with a bare
  io::Error (no path). detect_hypervisor now returns anyhow::Error with
  .with_context() naming the failing DMI path, so support bundles stay
  actionable.
- detect_hypervisor doc contradicted itself: it claimed None on
  unreadable DMI files, but the code always returns Err via ?. Trimmed
  the doc to only describe None as "read fine, unrecognized
  vendor/product".
- trident::virt duplicated the DMI/sysfs-probing pattern already
  established by osutils::virt::is_virtdeploy. Moved
  detect_hypervisor/is_virtual into osutils::virt (alongside
  is_virtdeploy) and deleted trident::virt, updating diagnostics.rs and
  tracestream.rs to call osutils::virt. bootentries.rs and rollback.rs
  already used osutils::virt::is_virtdeploy, so this removes the
  redundant module entirely.
